+++ src/galeforce/interfaces/dto.ts
@@ -1,67 +1,32 @@
-import { SummonerDTO } from './dto/summoner-v4/summoner';
-import { MatchDTO } from './dto/match-v4/match';
-import { MatchTimelineDTO } from './dto/match-v4/match-timeline';
-import { MatchlistDTO } from './dto/match-v4/matchlist';
-import { LeagueEntryDTO } from './dto/league/league-entry';
-import { ChampionMasteryDTO } from './dto/champion-mastery-v4/champion-mastery';
-import { PlatformDataDTO } from './dto/platform/platform-data';
-import { ChampionInfoDTO } from './dto/champion-v3/champion-info';
-import { PlayerDTO } from './dto/clash-v1/player';
-import { TeamDTO } from './dto/clash-v1/team';
-import { TournamentDTO } from './dto/clash-v1/tournament';
-import { CurrentGameInfoDTO } from './dto/spectator-v4/current-game-info';
-import { FeaturedGamesDTO } from './dto/spectator-v4/featured-games';
-import { LeagueListDTO } from './dto/league/league-list';
-import { AccountDTO } from './dto/account-v1/account';
-import { ActiveShardDTO } from './dto/account-v1/active-shard';
-import { TournamentCodeDTO } from './dto/tournament-v4/tournament-code';
-import { LobbyEventDTOWrapper } from './dto/tournament-v4/lobby-event';
-import { LorMatchDTO } from './dto/lor-match-v1/match';
-import { LorLeaderboardDTO } from './dto/lor-ranked-v1/leaderboard';
-import { TFTMatchDTO } from './dto/tft-match-v1/match';
-import { ValContentDTO } from './dto/val-content-v1/content';
-import { ValMatchDTO } from './dto/val-match-v1/match';
-import { ValMatchlistDTO } from './dto/val-match-v1/matchlist';
-import { ValRecentMatchesDTO } from './dto/val-match-v1/recent-matches';
-import { ValLeaderboardDTO } from './dto/val-ranked-v1/leaderboard';
-import { DataDragonRegionDTO } from './dto/data-dragon/regions';
-import { DataDragonChampionListDTO } from './dto/data-dragon/champion-list';
-import { DataDragonChampionDTO } from './dto/data-dragon/champion';
-import { DataDragonItemListDTO } from './dto/data-dragon/item-list';
-import { DataDragonSummonerListDTO } from './dto/data-dragon/summoner-spells';
-import { DataDragonProfileIconListDTO } from './dto/data-dragon/profile-icons';
-
-export {
-    SummonerDTO,
-    MatchDTO,
-    MatchTimelineDTO,
-    MatchlistDTO,
-    LeagueEntryDTO,
-    LeagueListDTO,
-    ChampionMasteryDTO,
-    PlatformDataDTO,
-    ChampionInfoDTO,
-    PlayerDTO,
-    TeamDTO,
-    TournamentDTO,
-    CurrentGameInfoDTO,
-    FeaturedGamesDTO,
-    AccountDTO,
-    ActiveShardDTO,
-    TournamentCodeDTO,
-    LobbyEventDTOWrapper,
-    LorMatchDTO,
-    LorLeaderboardDTO,
-    TFTMatchDTO,
-    ValContentDTO,
-    ValMatchDTO,
-    ValMatchlistDTO,
-    ValRecentMatchesDTO,
-    ValLeaderboardDTO,
-    DataDragonRegionDTO,
-    DataDragonChampionListDTO,
-    DataDragonChampionDTO,
-    DataDragonItemListDTO,
-    DataDragonSummonerListDTO,
-    DataDragonProfileIconListDTO,
-};
+export { SummonerDTO } from './dto/summoner-v4/summoner';
+export { MatchDTO } from './dto/match-v4/match';
+export { MatchTimelineDTO } from './dto/match-v4/match-timeline';
+export { MatchlistDTO } from './dto/match-v4/matchlist';
+export { LeagueEntryDTO } from './dto/league/league-entry';
+export { ChampionMasteryDTO } from './dto/champion-mastery-v4/champion-mastery';
+export { PlatformDataDTO } from './dto/platform/platform-data';
+export { ChampionInfoDTO } from './dto/champion-v3/champion-info';
+export { PlayerDTO } from './dto/clash-v1/player';
+export { TeamDTO } from './dto/clash-v1/team';
+export { TournamentDTO } from './dto/clash-v1/tournament';
+export { CurrentGameInfoDTO } from './dto/spectator-v4/current-game-info';
+export { FeaturedGamesDTO } from './dto/spectator-v4/featured-games';
+export { LeagueListDTO } from './dto/league/league-list';
+export { AccountDTO } from './dto/account-v1/account';
+export { ActiveShardDTO } from './dto/account-v1/active-shard';
+export { TournamentCodeDTO } from './dto/tournament-v4/tournament-code';
+export { LobbyEventDTOWrapper } from './dto/tournament-v4/lobby-event';
+export { LorMatchDTO } from './dto/lor-match-v1/match';
+export { LorLeaderboardDTO } from './dto/lor-ranked-v1/leaderboard';
+export { TFTMatchDTO } from './dto/tft-match-v1/match';
+export { ValContentDTO } from './dto/val-content-v1/content';
+export { ValMatchDTO } from './dto/val-match-v1/match';
+export { ValMatchlistDTO } from './dto/val-match-v1/matchlist';
+export { ValRecentMatchesDTO } from './dto/val-match-v1/recent-matches';
+export { ValLeaderboardDTO } from './dto/val-ranked-v1/leaderboard';
+export { DataDragonRegionDTO } from './dto/data-dragon/regions';
+export { DataDragonChampionListDTO } from './dto/data-dragon/champion-list';
+export { DataDragonChampionDTO } from './dto/data-dragon/champion';
+export { DataDragonItemListDTO } from './dto/data-dragon/item-list';
+export { DataDragonSummonerListDTO } from './dto/data-dragon/summoner-spells';
+export { DataDragonProfileIconListDTO } from './dto/data-dragon/profile-icons';